Do only formulae with brewl, brewo and brewU
mainly because, when `--formula` is not specified, they will work with
casks too at the same time. It's handy to have a way to work with
formulae and casks separately. For example, user might want to upgrade
all outdated and unpinned formulae, but not all outdated casks at the
same time (because there's no way to pin the casks you don't want to
upgrade when doing the batch operation).